requirements.txt
anyascii==0.3.2 ; python_version >= "3.11" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
beautifulsoup4==4.11.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
crispy-tailwind==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
django-appconf==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
django-compressor==4.3.1 ; python_version >= "3.11" and python_version < "4.0"
django-crispy-forms==2.0 ; python_version >= "3.11" and python_version < "4.0"
django-filter==22.1 ; python_version >= "3.11" and python_version < "4.0"
django-modelcluster==6.0 ; python_version >= "3.11" and python_version < "4.0"
django-permissionedforms==0.1 ; python_version >= "3.11" and python_version < "4.0"
django-taggit==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
django-treebeard==4.7 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.2 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
draftjs-exporter==2.1.7 ; python_version >= "3.11" and python_version < "4.0"
et-xmlfile==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
filetype==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
html5lib==1.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
l18n==2021.3 ; python_version >= "3.11" and python_version < "4.0"
openpyxl==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.11" and python_version < "4.0"
rcssmin==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rjsmin==1.2.1 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.4.1 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
telepath==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.0.3 ; python_version >= "3.11" and python_version < "4.0"
wagtail==5.0.2 ; python_version >= "3.11" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
willow==1.5 ; python_version >= "3.11" and python_version < "4.0"
